+++ types/src/front/assistant/actions/process.ts
@@ -91,3 +91,37 @@ export interface ProcessActionType extends BaseAction {
   step: number;
   type: "process_action";
 }
+
+/**
+ * Process Action Events
+ */
+
+// Event sent before the execution with the finalized params to be used.
+export type ProcessParamsEvent = {
+  type: "process_params";
+  created: number;
+  configurationId: string;
+  messageId: string;
+  dataSources: DataSourceConfiguration[];
+  action: ProcessActionType;
+};
+
+export type ProcessErrorEvent = {
+  type: "process_error";
+  created: number;
+  configurationId: string;
+  messageId: string;
+  error: {
+    code: string;
+    message: string;
+  };
+};
+
+export type ProcessSuccessEvent = {
+  type: "process_success";
+  created: number;
+  configurationId: string;
+  messageId: string;
+  action: ProcessActionType;
+};
+
