+++ types/src/front/assistant/actions/tables_query.ts
@@ -67,3 +67,42 @@ export function getTablesQueryResultsFileAttachment({
 
   return `${attachment}>\n${resultsFileSnippet}\n</file>`;
 }
+
+/**
+ * TablesQuey Events
+ */
+
+export type TablesQueryErrorEvent = {
+  type: "tables_query_error";
+  created: number;
+  configurationId: string;
+  messageId: string;
+  error: {
+    code: "tables_query_error" | "too_many_result_rows";
+    message: string;
+  };
+};
+
+export type TablesQueryStartedEvent = {
+  type: "tables_query_started";
+  created: number;
+  configurationId: string;
+  messageId: string;
+  action: TablesQueryActionType;
+};
+
+export type TablesQueryModelOutputEvent = {
+  type: "tables_query_model_output";
+  created: number;
+  configurationId: string;
+  messageId: string;
+  action: TablesQueryActionType;
+};
+
+export type TablesQueryOutputEvent = {
+  type: "tables_query_output";
+  created: number;
+  configurationId: string;
+  messageId: string;
+  action: TablesQueryActionType;
+};
