+++ types/src/front/assistant/generation.ts
