--- types/src/front/lib/api/assistant/actions/dust_app_run.ts
@@ -1,41 +1,0 @@
-/**
- * Action execution.
- */
-
-import { DustAppRunActionType } from "../../../../../front/assistant/actions/dust_app_run";
-
-// Event sent before the execution of a dust app run with the finalized params to be used.
-export type DustAppRunParamsEvent = {
-  type: "dust_app_run_params";
-  created: number;
-  configurationId: string;
-  messageId: string;
-  action: DustAppRunActionType;
-};
-
-export type DustAppRunErrorEvent = {
-  type: "dust_app_run_error";
-  created: number;
-  configurationId: string;
-  messageId: string;
-  error: {
-    code: string;
-    message: string;
-  };
-};
-
-export type DustAppRunBlockEvent = {
-  type: "dust_app_run_block";
-  created: number;
-  configurationId: string;
-  messageId: string;
-  action: DustAppRunActionType;
-};
-
-export type DustAppRunSuccessEvent = {
-  type: "dust_app_run_success";
-  created: number;
-  configurationId: string;
-  messageId: string;
-  action: DustAppRunActionType;
-};
