+++ types/src/index.ts
@@ -24,6 +24,7 @@ export * from "./front/assistant/actions/browse";
 export * from "./front/assistant/actions/conversation/list_files";
 export * from "./front/assistant/actions/dust_app_run";
 export * from "./front/assistant/actions/guards";
+export * from "./front/assistant/actions/index";
 export * from "./front/assistant/actions/process";
 export * from "./front/assistant/actions/retrieval";
 export * from "./front/assistant/actions/tables_query";
@@ -32,6 +33,7 @@ export * from "./front/assistant/agent";
 export * from "./front/assistant/avatar";
 export * from "./front/assistant/builder";
 export * from "./front/assistant/conversation";
+export * from "./front/assistant/generation";
 export * from "./front/assistant/pubsub";
 export * from "./front/assistant/templates";
 export * from "./front/assistant/visualization";
@@ -45,14 +47,6 @@ export * from "./front/files";
 export * from "./front/groups";
 export * from "./front/key";
 export * from "./front/lib/actions/types";
-export * from "./front/lib/api/assistant/actions/browse";
-export * from "./front/lib/api/assistant/actions/dust_app_run";
-export * from "./front/lib/api/assistant/actions/index";
-export * from "./front/lib/api/assistant/actions/process";
-export * from "./front/lib/api/assistant/actions/retrieval";
-export * from "./front/lib/api/assistant/actions/tables_query";
-export * from "./front/lib/api/assistant/actions/websearch";
-export * from "./front/lib/api/assistant/generation";
 export * from "./front/lib/api/credentials";
 export * from "./front/lib/assistant";
 export * from "./front/lib/connectors_api";
